swapskill/skill-swap-backend/requirements.txt
Django==5.2.3
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
channels==4.2.2
daphne==4.2.1
Pillow==11.3.0
python-dotenv==1.1.1
psycopg2-binary==2.9.10
dj-database-url==3.0.1
gunicorn==23.0.0
whitenoise==6.9.0
django-allauth==65.8.1
social-auth-app-django==5.4.3
requests-oauthlib==2.0.0
twilio==8.10.0
requests==2.32.4
python-decouple==3.8
django-ratelimit==4.1.0
redis==5.0.1
celery==5.3.4
phonenumbers==8.13.26
bleach==6.1.0

PyJWT==2.8.0
cryptography==42.0.8

cffi==1.16.0
